+++ test/interactive/interactive_test3.cpp
@@ -67,6 +67,27 @@ SystemError: glob expansion is canceled, caused by `%s'
   ASSERT_NO_FATAL_FAILURE(this->sendLineAndWait("exit", 1));
 }
 
+TEST_F(InteractiveTest, expand_ctrlc4) {
+  this->invoke("--quiet", "--norc");
+
+  ASSERT_NO_FATAL_FAILURE(this->expect(PROMPT));
+  this->sendLine("source? /*//*//*/*//*/*//*/*/*//**/?!/%/*/*/*/s*/../*/../*");
+  ASSERT_NO_FATAL_FAILURE(
+      this->expect(PROMPT + "source? /*//*//*/*//*/*//*/*/*//**/?!/%/*/*/*/s*/../*/../*\n"));
+  this->send(CTRL_C);
+
+  std::string err = R"([semantic error] glob or brace expansion canceled
+ --> (stdin):1:9
+source? /*//*//*/*//*/*//*/*/*//**/?!/%/*/*/*/s*/../*/../*
+        ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+)";
+
+  ASSERT_NO_FATAL_FAILURE(this->expect(promptAfterCtrlC(PROMPT), err));
+
+  // last exit status is 0 (does not update $?)
+  ASSERT_NO_FATAL_FAILURE(this->sendLineAndWait("exit", 0));
+}
+
 TEST_F(InteractiveTest, wait_ctrlc1) {
   this->invoke("--quiet", "--norc");