+++ CHANGELOG.md
@@ -62,6 +62,7 @@
     - Option type within string interpolation / parameter expansion
     - nested job operator
 - reimplement ``Map`` object. now always preserve insertion order
+- cancel glob or brace expansion in source statement
 
 #### Builtin